+++ src/apps/ava_import_ldap/admin.py
@@ -1,7 +1,7 @@
 from django.contrib import admin
 
 # Register your models here.
-from apps.ava_core_ldap.models import ActiveDirectoryUser, ActiveDirectoryGroup,LDAPConfiguration
+from apps.ava_import_ldap.models import ActiveDirectoryUser, ActiveDirectoryGroup,LDAPConfiguration
 
 admin.site.register(ActiveDirectoryUser)
 admin.site.register(ActiveDirectoryGroup)
