+++ src/apps/ava_import_ldap/templates/ldap/ActiveDirectoryGroup.html
