+++ src/apps/ava_vis_graph/models.py
@@ -2,8 +2,8 @@ from apps.ava_core.models import TimeStampedModel
 
 #class LDAPEdges(TimeStampedModel)
 #    name = models.CharField(max_length=50)
-#    source = models.ForeignKey('apps.ava_core_ldap.ActiveDirectoryUser')
-#    target = models.ForeignKey('apps.ava_core_ldap.ActiveDirectoryGroup')
+#    source = models.ForeignKey('apps.ava_import_ldap.ActiveDirectoryUser')
+#    target = models.ForeignKey('apps.ava_import_ldap.ActiveDirectoryGroup')
 
  #   def __unicode__(self):
  #       return self.name or u''
