+++ src/settings.py
@@ -96,8 +96,8 @@ LOCAL_APPS = (
     'apps.ava_core_group',
     'apps.ava_core_auth',
     'apps.ava_core_project',
-    'apps.ava_core_ldap',
-   # 'apps.ava_core_google_apps',
+    'apps.ava_import_ldap',
+   # 'apps.ava_import_google',
     'apps.ava_core_identity',
     'apps.ava_vis_graph',
     'apps.ava_test',
