+++ src/urls.py
@@ -15,7 +15,7 @@ urlpatterns = patterns('',
     url(r'^project/', include('apps.ava_core_project.urls')),
     url(r'^accounts/', include('apps.ava_core_auth.urls')),
     url(r'^group/', include('apps.ava_core_group.urls')),
-    url(r'^ldap/', include('apps.ava_core_ldap.urls')),
+    url(r'^ldap/', include('apps.ava_import_ldap.urls')),
     url(r'^people/', include('apps.ava_core_identity.urls')),
     url(r'^graph/', include('apps.ava_vis_graph.urls')),
     url(r'^test/', include('apps.ava_test.urls')),
